init method called

+++ Swift_Payroll_Project/Employee.swift
@@ -26,8 +26,8 @@ func calcBirthYear() -> Int{
     //returning result
     return currYear - age
 }
-func calcEarnings() -> Int{
-        return 1000;
+func calcEarnings() -> Double{
+        return 1000.00;
     }
 func printMyData() {
         print("Name   : \(name)")
